requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
hyperliquid-python-sdk>=0.1.0
eth-account>=0.9.0
web3>=6.0.0
python-dotenv>=1.0.0
# Added for CoinGecko API and Telegram
pycoingecko>=3.2.0 
python-telegram-bot>=20.0
# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0 